test: update test naming, use pytest-cases for mp6 tests, run linter
