+++ autotest/test_modflow.py
@@ -10,9 +10,7 @@ from autotest.conftest import (
     excludes_platform,
     get_example_data_path,
     requires_exe,
-    requires_pkg,
 )
-from matplotlib import pyplot as plt
 
 from flopy.discretization import StructuredGrid
 from flopy.mf6 import MFSimulation
@@ -31,9 +29,8 @@ from flopy.modflow import (
     ModflowWel,
 )
 from flopy.mt3d import Mt3dBtn, Mt3dms
-from flopy.plot import PlotMapView
 from flopy.seawat import Seawat
-from flopy.utils import EndpointFile, PathlineFile, TimeseriesFile, Util2d
+from flopy.utils import Util2d
 
 
 @pytest.fixture
@@ -942,140 +939,6 @@ def test_rchload(tmpdir):
     assert np.allclose(a1, a2)
     a2 = m2l.rch.rech[1].array
     assert np.allclose(a1, a2)
-
-
-@requires_pkg("pandas")
-def test_mp5_load(tmpdir, example_data_path):
-    # load the base freyberg model
-    freyberg_ws = example_data_path / "freyberg"
-    # load the modflow files for model map
-    m = Modflow.load(
-        "freyberg.nam",
-        model_ws=str(freyberg_ws),
-        check=False,
-        verbose=True,
-        forgive=False,
-    )
-
-    # load the pathline data
-    pthobj = PathlineFile(str(example_data_path / "mp5" / "m.ptl"))
-
-    # load endpoint data
-    fpth = str(example_data_path / "mp5" / "m.ept")
-    endobj = EndpointFile(fpth, verbose=True)
-
-    # determine version
-    ver = pthobj.version
-    assert ver == 5, f"{fpth} is not a MODPATH version 5 pathline file"
-
-    # read all of the pathline and endpoint data
-    plines = pthobj.get_alldata()
-    epts = endobj.get_alldata()
-
-    # determine the number of particles in the pathline file
-    nptl = pthobj.nid.shape[0]
-    assert nptl == 64, "number of MODPATH 5 particles does not equal 64"
-
-    hsv = plt.get_cmap("hsv")
-    colors = hsv(np.linspace(0, 1.0, nptl))
-
-    # plot the pathlines one pathline at a time
-    mm = PlotMapView(model=m)
-    for n in pthobj.nid:
-        p = pthobj.get_data(partid=n)
-        e = endobj.get_data(partid=n)
-        try:
-            mm.plot_pathline(p, colors=colors[n], layer="all")
-        except:
-            assert False, f'could not plot pathline {n + 1} with layer="all"'
-        try:
-            mm.plot_endpoint(e)
-        except:
-            assert False, f'could not plot endpoint {n + 1} with layer="all"'
-
-    # plot the grid and ibound array
-    mm.plot_grid(lw=0.5)
-    mm.plot_ibound()
-
-    fpth = os.path.join(str(tmpdir), "mp5.pathline.png")
-    plt.savefig(fpth, dpi=300)
-    plt.close()
-
-
-@requires_pkg("pandas")
-def test_mp5_timeseries_load(example_data_path):
-    pth = str(example_data_path / "mp5")
-    files = [
-        os.path.join(pth, name)
-        for name in sorted(os.listdir(pth))
-        if ".timeseries" in name
-    ]
-    for file in files:
-        print(file)
-        eval_timeseries(file)
-
-
-@requires_pkg("pandas")
-def test_mp6_timeseries_load(example_data_path):
-    pth = str(example_data_path / "mp5")
-    files = [
-        os.path.join(pth, name)
-        for name in sorted(os.listdir(pth))
-        if ".timeseries" in name
-    ]
-    for file in files:
-        print(file)
-        eval_timeseries(file)
-
-
-def eval_timeseries(file):
-    ts = TimeseriesFile(file)
-    assert isinstance(ts, TimeseriesFile), (
-        f"{os.path.basename(file)} " "is not an instance of TimeseriesFile"
-    )
-
-    # get the all of the data
-    tsd = ts.get_alldata()
-    assert (
-        len(tsd) > 0
-    ), f"could not load data using get_alldata() from {os.path.basename(file)}."
-
-    # get the data for the last particleid
-    partid = ts.get_maxid()
-    assert partid is not None, (
-        "could not get maximum particleid using get_maxid() from "
-        f"{os.path.basename(file)}."
-    )
-
-    tsd = ts.get_data(partid=partid)
-    assert tsd.shape[0] > 0, (
-        f"could not load data for particleid {partid} using get_data() from "
-        f"{os.path.basename(file)}. Maximum partid = {ts.get_maxid()}."
-    )
-
-    timemax = ts.get_maxtime() / 2.0
-    assert timemax is not None, (
-        "could not get maximum time using get_maxtime() from "
-        f"{os.path.basename(file)}."
-    )
-
-    tsd = ts.get_alldata(totim=timemax)
-    assert len(tsd) > 0, (
-        f"could not load data for totim>={timemax} using get_alldata() from "
-        f"{os.path.basename(file)}. Maximum totim = {ts.get_maxtime()}."
-    )
-
-    timemax = ts.get_maxtime()
-    assert timemax is not None, (
-        "could not get maximum time using get_maxtime() from "
-        f"{os.path.basename(file)}."
-    )
-
-    tsd = ts.get_alldata(totim=timemax, ge=False)
-    assert len(tsd) > 0, (
-        f"could not load data for totim<={timemax} using get_alldata() from "
-        f"{os.path.basename(file)}. Maximum totim = {ts.get_maxtime()}."
-    )
 
 
 def test_default_oc_stress_period_data(tmpdir):
